Add BackButton component and integrate it into Login and Register pages
