+++ frontend/app/(auth)/login/page.tsx
@@ -10,6 +10,7 @@ import { Button } from "@/components/ui/button";
 import Input from "@/components/ui/Input";
 import { getApiErrorMessage } from "@/lib/api/error";
 import { loginUser } from "@/lib/api/services/auth-service";
+import BackButton from "@/components/ui/back-button";
 
 export default function LoginPage() {
   const router = useRouter();
@@ -53,6 +54,9 @@ export default function LoginPage() {
 
   return (
     <main className="mx-auto w-full max-w-md">
+      <div className="mb-6">
+        <BackButton />
+      </div>
       <div className="mb-10 text-center">
         <div className="mx-auto mb-5 flex items-center justify-center">
           <Image
