+++ frontend/app/leaderboard/_components/leaderboard-back-button.tsx
@@ -1,31 +1,14 @@
 "use client";
 
-import { ArrowLeft } from "lucide-react";
-import { useRouter } from "next/navigation";
-import { Button } from "@/components/ui/button";
 import type { LeaderboardSource } from "../_lib/leaderboard-data";
+import BackButton from "@/components/ui/back-button";
 
 type LeaderboardBackButtonProps = {
   source: LeaderboardSource;
+  className?: string;
 };
 
-export default function LeaderboardBackButton({
-  source,
-}: LeaderboardBackButtonProps) {
-  const router = useRouter();
-
-  function handleBack() {
-    router.push(source === "home" ? "/" : "/dashboard");
-  }
-
-  return (
-    <Button
-      variant="ghost"
-      className="mb-6 inline-flex gap-2 text-slate-600 hover:text-slate-900"
-      onClick={handleBack}
-    >
-      <ArrowLeft className="h-4 w-4" />
-      <span className="text-sm font-medium">Back</span>
-    </Button>
-  );
+export default function LeaderboardBackButton({ source, className }: LeaderboardBackButtonProps) {
+  const to = source === "home" ? "/" : "/dashboard";
+  return <BackButton to={to} className={`mb-6 ${className ?? ""}`} />;
 }
